--- mgmt.version
@@ -1,3 +1,0 @@
-{
-    "version": "25.03.25.1"
-}
