+++ mgmt/mgmt.version
@@ -1,0 +1,3 @@
+{
+    "version": "25.07.24.0"
+}
