+++ mgmt/mgmt_LockScreen.py
@@ -23,9 +23,10 @@ import time
 import psutil
 import traceback
 
-from color import p
 
-import util
+from common.color import p
+
+from common import util
 from mgmt_UserAccounts import UserAccounts
 from mgmt_ProcessManagement import ProcessManagement
 from mgmt_Computer import Computer
