+++ mgmt/mgmt_SystemTime.py
@@ -5,11 +5,12 @@ import json
 
 import wmi
 
-import util
+from common import util
 from mgmt_RegistrySettings import RegistrySettings
 from mgmt_ProcessManagement import ProcessManagement
 
-from color import p
+
+from common.color import p
 from p_state import p_state
 
 class SystemTime:
