+++ mgmt/rc/STENCIL.TTF
