+++ mgmt/rc/ad_gpo/{6B3E1AFF-B4F9-4DD4-8E93-EDF58C4930E1}/DomainSysvol/GPO/User/registry.pol
