+++ mgmt/rc/gpo/{4C67F01C-820A-4D17-81C0-0FB464EF65D0}/DomainSysvol/GPO/Machine/microsoft/windows nt/Audit/audit.csv
