+++ mgmt/rc/install_vc_runtimes.cmd
