+++ mgmt/rc/magic.py
