+++ mgmt/rc/upgrade_ope.cmd
